requirements-dev.txt
asgiref==3.5.2
certifi==2022.9.14
chardet==4.0.0
dj-database-url==1.0.0
dj-static==0.0.6
Django==3.2.6
django-heroku==0.3.1
djangorestframework==3.13.1
gitdb==4.0.9
GitPython==3.1.27
idna==2.10
psycopg2==2.9.3
PyJWT==2.5.0
pyotp==2.7.0
pytz==2022.2.1
requests==2.25.1
smmap==5.0.0
sqlparse==0.4.2
static3==0.7.0
twilio==7.14.0
typing_extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
